# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Traveloop API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips/{tripId}/stops:
    get:
      operationId: listStops
      tags: [stops]
      summary: List stops for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stop"
    post:
      operationId: createStop
      tags: [stops]
      summary: Add a stop to a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StopInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stop"
  /trips/{tripId}/stops/reorder:
    put:
      operationId: reorderStops
      tags: [stops]
      summary: Reorder stops in a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StopReorder"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /trips/{tripId}/stops/{stopId}:
    patch:
      operationId: updateStop
      tags: [stops]
      summary: Update a stop
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: stopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StopUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stop"
    delete:
      operationId: deleteStop
      tags: [stops]
      summary: Remove a stop from a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: stopId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Activities on a stop
...
  /trips/{tripId}/packing:
    get:
      operationId: listPackingItems
      tags: [packing]
      summary: List packing items for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PackingItem"
    post:
      operationId: createPackingItem
      tags: [packing]
      summary: Add a packing item
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackingItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackingItem"
  /trips/{tripId}/packing/{itemId}:
    patch:
      operationId: updatePackingItem
      tags: [packing]
      summary: Update a packing item (toggle packed, rename, etc.)
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackingItemUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackingItem"
    delete:
      operationId: deletePackingItem
      tags: [packing]
      summary: Delete a packing item
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Notes
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Stop:
      type: object
      required: [id, tripId, cityId, cityName, country, startDate, endDate, order]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        cityId:
          type: integer
        cityName:
          type: string
        country:
          type: string
        cityImageUrl:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        order:
          type: integer
        activities:
          type: array
          items:
            $ref: "#/components/schemas/StopActivity"
    StopInput:
      type: object
      required: [cityId, startDate, endDate]
      properties:
        cityId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
    StopUpdate:
      type: object
      properties:
        cityId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
    StopReorder:
      type: object
      required: [order]
      properties:
        order:
          type: array
          items:
            type: integer
...
    StopActivity:
      type: object
      required: [id, stopId, name, category]
      properties:
        id:
          type: integer
        stopId:
          type: integer
        activityCatalogId:
          type: ["integer", "null"]
        name:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        cost:
          type: ["number", "null"]
        duration:
          type: ["number", "null"]
        scheduledTime:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
...
    PackingItem:
      type: object
      required: [id, tripId, name, category, isPacked, createdAt]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        name:
          type: string
        category:
          type: string
        isPacked:
          type: boolean
        createdAt:
          type: string
    PackingItemInput:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
    PackingItemUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        isPacked:
          type: boolean
...