# lib/api-zod/src/generated/types/stop.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Traveloop API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StopActivity } from "./stopActivity";

export interface Stop {
  id: number;
  tripId: number;
  cityId: number;
  cityName: string;
  country: string;
  /** @nullable */
  cityImageUrl?: string | null;
  startDate: string;
  endDate: string;
  order: number;
  activities?: StopActivity[];
}