# lib/api-zod/src/generated/types/stopUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Traveloop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StopUpdate {
  cityId?: number;
  startDate?: string;
  endDate?: string;
}